+++ config/imager.php
@@ -1,0 +1,10 @@
+<?php
+return [
+    'fillTransforms ' => false,
+    'fillInterval' => 256,
+
+    'resizeFilter ' => 'lanczos',
+    'webpQuality ' => 82,
+    'jpegQuality  ' => 82,
+    'useForNativeTransforms ' => false,
+];