+++ migrations/m220131_103453_create_guide_entries.php
@@ -29,7 +29,7 @@ class m220131_103453_create_guide_entries extends Migration
             ['title' => 'Inhalt', 'slug' => 'content'],
             ['title' => 'Dateien', 'slug' => 'assets'],
             ['title' => 'Content Builder', 'slug' => 'contentbuilder'],
-            ['title' => 'Markdown', 'slug' => 'markdown'],
+            ['title' => 'Blöcke', 'slug' => 'blocks'],
             ['title' => 'Sektionen', 'slug' => 'sections'],
             ['title' => 'Artikel', 'slug' => 'article'],
             ['title' => 'Seite', 'slug' => 'page'],
@@ -49,6 +49,63 @@ class m220131_103453_create_guide_entries extends Migration
                 echo "Error saving guide entry #{guide['title']}  \n";
             }
         }
+
+        $this->setParent('content', ['contentbuilder', 'blocks']);
+        $this->setParent('sections', ['article', 'page', 'legal']);
+
+        $this->setIncludeGuides(['article', 'page', 'legal'], ['contentBuilder', 'blocks']);
+    }
+
+    protected function setParent($parentSlug, $childrenSlugs)
+    {
+        $parent = $this->getGuideBySlug($parentSlug);
+        if (!$parent) {
+            return;
+        }
+
+        foreach ($childrenSlugs as $childSlug) {
+            /** @var Entry $child */
+            $entry = $this->getGuideBySlug($childSlug);
+
+            if (!$entry) {
+                continue;
+            }
+            $entry->newParentId = $parent->id;
+            Craft::$app->elements->saveElement($entry);
+        }
+    }
+
+    protected function setIncludeGuides($sectionSlugs, $includeSlugs)
+    {
+        $includeIds = [];
+        foreach ($includeSlugs as $includeSlug) {
+            $entry = $this->getGuideBySlug($includeSlug);
+
+            if (!$entry) {
+                continue;
+            }
+            $includeIds[] = $entry->id;
+        }
+
+        if (!$includeIds) {
+            return;
+        }
+
+        foreach ($sectionSlugs as $sectionSlug) {
+            $entry = $this->getGuideBySlug($sectionSlug);
+
+            if (!$entry) {
+                continue;
+            }
+
+            $entry->setFieldValue('includeGuides', $includeIds);
+            Craft::$app->elements->saveElement($entry);
+        }
+    }
+
+    protected function getGuideBySlug($slug)
+    {
+        return Entry::find()->section('guide')->slug($slug)->one();
     }
 
     /**
@@ -56,7 +113,7 @@ class m220131_103453_create_guide_entries extends Migration
      */
     public function safeDown()
     {
-        echo "m220131_103453_create_guide_entries cannot be reverted.\n";
+        echo "There is nothing to revert.\n";
         return true;
     }
 }