+++ src/index.js
@@ -7,9 +7,11 @@ window.Alpine = Alpine
 
 import focus from '@alpinejs/focus'
 import collapse from '@alpinejs/collapse'
+import intersect from '@alpinejs/intersect'
 
 Alpine.plugin(focus)
 Alpine.plugin(collapse)
+Alpine.plugin(intersect)
 
 Alpine.store('consent', Consent)
 Alpine.store('modalOpen', false)